fix: remove init command
