+++ packages/graphback-cli/src/helpers/index.ts
@@ -1,6 +1,5 @@
 export * from './common'
 export * from './db'
 export * from './generate'
-export * from './init'
 export * from './config'
 export * from './transformOpenApiSpec'
